* src/gsad_omp.h: Updated according to last commit.
git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/gsa@6239 423fd1db-d629-0410-8442-d21db03e70f4

+++ src/gsad_omp.h
@@ -62,7 +62,7 @@ char * create_lsc_credential_omp (credentials_t *, char *, char *,
 char * delete_lsc_credential_omp (credentials_t *, const char *);
 
 char * get_agents_omp (credentials_t *, const char *, const char *,
-                       unsigned int *, const char *, const char *);
+                       gsize *, const char *, const char *);
 char * create_agent_omp (credentials_t *, const char *, const char *,
                          const char *, int,
                          const char *, int,
